+++ internal/infrastructure/controllers/user_find_by_id_test.go
@@ -54,7 +54,7 @@ func TestFindUserByIDController_Handle(t *testing.T) {
 			testName:     "success - user found",
 			routeParamID: "1",
 			setupMock: func(m *mockApp.MockFindUserByIDUseCase) {
-				m.EXPECT().Execute(mock.Anything, 1).Return(validUser, nil)
+				m.EXPECT().Execute(mock.Anything, int64(1)).Return(validUser, nil)
 			},
 			expectedStatus: http.StatusOK,
 		},
@@ -83,7 +83,7 @@ func TestFindUserByIDController_Handle(t *testing.T) {
 			testName:     "fail - user not found",
 			routeParamID: "1",
 			setupMock: func(m *mockApp.MockFindUserByIDUseCase) {
-				m.EXPECT().Execute(mock.Anything, 1).Return(nil, domain.ErrUserNotFound)
+				m.EXPECT().Execute(mock.Anything, int64(1)).Return(nil, domain.ErrUserNotFound)
 			},
 			expectedStatus: http.StatusNotFound,
 			expectedBody:   domain.ErrUserNotFound.Error(),
@@ -92,7 +92,7 @@ func TestFindUserByIDController_Handle(t *testing.T) {
 			testName:     "fail - internal server error from usecase",
 			routeParamID: "1",
 			setupMock: func(m *mockApp.MockFindUserByIDUseCase) {
-				m.EXPECT().Execute(mock.Anything, 1).Return(nil, errors.New("database connection lost"))
+				m.EXPECT().Execute(mock.Anything, int64(1)).Return(nil, errors.New("database connection lost"))
 			},
 			expectedStatus: http.StatusInternalServerError,
 			expectedBody:   "internal server error",
